requirements.txt
flask
redis
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
supabase
flask-cors
google-generativeai